split anari_test_scenes into different categories
