+++ libs/anari_test_scenes/scenes/demo/cornell_box.cpp
