+++ libs/anari_test_scenes/scenes/demo/gravity_spheres_volume.cpp
