+++ libs/anari_test_scenes/scenes/file/tiny_obj_loader.h
