+++ libs/anari_test_scenes/scenes/test/attributes.cpp
