+++ libs/anari_test_scenes/scenes/test/instanced_cubes.cpp
