+++ libs/anari_test_scenes/scenes/test/pbr_spheres.cpp
