+++ libs/anari_test_scenes/scenes/test/random_cylinders.cpp
