+++ libs/anari_test_scenes/scenes/test/random_spheres.cpp
