+++ libs/anari_test_scenes/scenes/test/textured_cube.cpp
