+++ libs/anari_test_scenes/scenes/test/textured_cube.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "scene.h"
+#include "../scene.h"
 
 namespace anari {
 namespace scenes {
